dotCMS/core#24220 limited-users-should-not-edit-favorite-pages - PR feedback
